requirements.txt
# Data manipulation and analysis
pandas==2.1.3
numpy==1.26.2

# Data generation
Faker==20.1.0

# Machine Learning and Preprocessing
scikit-learn==1.3.2

# FastAPI and web server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Data validation
pydantic==2.5.2
email-validator==2.1.0.post1

# Testing
pytest==7.4.3
httpx==0.25.2

# Development tools
python-dotenv==1.0.0
black==23.11.0
flake8==6.1.0

pydantic_settings
uvicorn